test(branch): move all tests related to Branch-declared methods to the correct namespace
