+++ Tests/YDotNet.Tests.Unit/Branches/MapUnobserveDeepTests.cs
@@ -1,32 +1,33 @@
 using NUnit.Framework;
 using YDotNet.Document;
+using YDotNet.Document.Cells;
 
-namespace YDotNet.Tests.Unit.XmlTexts;
+namespace YDotNet.Tests.Unit.Branches;
 
-public class UnobserveDeepTests
+public class MapUnobserveDeepTests
 {
     [Test]
-    public void TriggersWhenChangedUntilUnobserved()
+    public void TriggersWhenMapChangedUntilUnobserved()
     {
         // Arrange
         var doc = new Doc();
-        var xmlText = doc.XmlText("xml-text");
+        var map = doc.Map("map");
         var called = 0;
-        var subscription = xmlText.ObserveDeep(_ => called++);
+        var subscription = map.ObserveDeep(_ => called++);
 
         // Act
         var transaction = doc.WriteTransaction();
-        xmlText.Insert(transaction, index: 0, "World");
+        map.Insert(transaction, "value1", Input.Long(value: 2469L));
         transaction.Commit();
 
         // Assert
         Assert.That(called, Is.EqualTo(expected: 1));
 
         // Act
-        xmlText.UnobserveDeep(subscription);
+        map.UnobserveDeep(subscription);
 
         transaction = doc.WriteTransaction();
-        xmlText.Insert(transaction, index: 0, "Hello, ");
+        map.Insert(transaction, "value2", Input.Long(value: -420L));
         transaction.Commit();
 
         // Assert
